Menu script, WIP
