+++ src/assets/scripts/Pinecone/Menu/index.js
@@ -1,80 +1,3 @@
-'use strict';
-
-/**
- * Accordion class.
- */
-class Accordion {
-	/**
-	 * Constructor.
-	 *
-	 * @param {DomNode} container
-	 * @param {Object} options
-	 */
-	constructor( container, options ) {
-		this.container = container;
-		this.config = {
-			...{
-				headingSelector: '.accordion__heading',
-				paneSelector: '.accordion__pane',
-				controlSelector: '.accordion__control'
-			},
-			...options
-		};
-		this.panes = this.container.querySelectorAll( this.config.paneSelector );
-		Array.prototype.forEach.call( this.panes, pane => {
-			this.initPane( pane );
-		} );
-		this.handleClicks = this.handleClicks.bind( this );
-		this.addEventListeners();
-	}
-
-	/**
-	 * Initialize a pane within the accordion.
-	 *
-	 * @param {DomNode} pane
-	 */
-	initPane( pane ) {
-		const heading = pane.querySelector( this.config.headingSelector );
-		const ctrl = document.createElement( 'button' );
-		ctrl.setAttribute( 'class', 'accordion__control' );
-		ctrl.setAttribute( 'aria-expanded', 'false' );
-		ctrl.setAttribute( 'type', 'button' );
-		ctrl.innerHTML = `
-			${heading.textContent}
-			<svg class="icon icon--chevron-down" viewBox="0 0 20 20" aria-hidden="true" focusable="false" xmlns="http://www.w3.org/2000/svg"><path id="chevron-down" d="m10 15a1 1 0 0 1 -.71-.29l-5-5a1 1 0 0 1 1.42-1.42l4.29 4.3 4.29-4.3a1 1 0 0 1 1.42 1.42l-5 5a1 1 0 0 1 -.71.29z" fill="currentColor"/></svg>
-		`;
-		heading.parentNode.insertBefore( ctrl, heading.nextElementSibling );
-		heading.parentNode.removeChild( heading );
-	}
-
-	/**
-	 *
-	 * @param {Event} event
-	 */
-	handleClicks( event ) {
-		if ( ! event.target.closest( this.config.controlSelector ) ) return;
-
-		const ctrl = event.target.closest( this.config.controlSelector );
-		const expanded = 'true' === ctrl.getAttribute( 'aria-expanded' ) || false;
-		Array.prototype.forEach.call( this.container.querySelectorAll( '.accordion__control' ), el => {
-			el.setAttribute( 'aria-expanded', false );
-		} );
-		ctrl.setAttribute( 'aria-expanded', !expanded );
-		if ( expanded ) {
-			ctrl.parentNode.classList.remove( 'accordion__pane--expanded' );
-		} else {
-			ctrl.parentNode.classList.add( 'accordion__pane--expanded' );
-		}
-	}
-
-	/**
-	 * Add click event listeners.
-	 */
-	addEventListeners() {
-		this.container.addEventListener( 'click', this.handleClicks, false );
-	}
-}
-
 /**
  * Menu class.
  */
@@ -173,10 +96,10 @@ class Menu {
 		const subMenu = el.closest( this.config.childMenuSelector );
 		// Last focus was on an item in a submenu; maybe we need to close it
 		if ( subMenu ) {
-			const items = subMenu.querySelectorAll( 'li' );
-			const itemCount = items.length;
-			const item = el.parentNode;
-			console.log( Array.prototype.indexOf.call( items, item ) );
+			// const items = subMenu.querySelectorAll( 'li' );
+			// const itemCount = items.length;
+			// const item = el.parentNode;
+			// console.log( Array.prototype.indexOf.call( items, item ) );
 			// const parentMenu = subMenu.closest( this.config.parentMenuSelector );
 			// Focus is still within the parent menu; we should leave it open.
 			// if ( document.activeElement.closest( this.config.parentMenuSelector ) === parentMenu ) return false;
@@ -198,6 +121,8 @@ class Menu {
 		document.addEventListener( 'focusout', this.handleDropdownFocusOut, false );
 	}
 }
+
+export default Menu;
 
 /*
 export function menu() {
@@ -258,7 +183,3 @@ export function menu() {
 	}
 }
 */
-
-var index = { Accordion, Menu };
-
-module.exports = index;
