+++ src/assets/scripts/Pinecone/index.js
@@ -1,3 +1,4 @@
 import Accordion from './Accordion/index.js';
+import Menu from './Menu/index.js';
 
-export default { Accordion };
+export default { Accordion, Menu };
