+++ src/assets/scripts/pinecone.js
@@ -1,6 +1,16 @@
 /* global Pinecone */
 
+const menu = document.querySelector( '.menu' );
+const menuToggle = document.querySelector( '.menu-toggle' );
+
+if ( menu && menuToggle ) {
+	new Pinecone.Menu( menu, menuToggle );
+}
+
 const accordions = document.querySelectorAll( '.accordion' );
-Array.prototype.forEach.call( accordions, accordion => {
-	new Pinecone.Accordion( accordion );
-} );
+
+if ( accordions ) {
+	Array.prototype.forEach.call( accordions, accordion => {
+		new Pinecone.Accordion( accordion );
+	} );
+}
