+++ src/components/02-molecules/00-menu/menu.config.js
@@ -3,7 +3,6 @@ module.exports = {
 	status: 'wip',
 	context: {
 		inBanner: true,
-		scriptCall: 'window.pinecone.menu();',
 		menuTitle: 'Menu',
 		menuItems:  {
 			'menu-item-1': {
