+++ src/components/02-molecules/00-menu/menu.njk
@@ -1,7 +1,7 @@
 <nav aria-labelledby="menu-primary-label">
 	<button class="menu-toggle" aria-expanded="false">
-		{% render '@svg', {svg:'menu', class:'open'} %}
-		{% render '@svg', {svg:'close'} %}
+		{% render '@svg', {svg:'menu', class:'open'}, true %}
+		{% render '@svg', {svg:'close'}, true %}
 		<span id="menu-primary-label" class="menu-toggle__label">{{ menuTitle }}</span>
 	</button>
 	<ul class="menu{% if modifier %} menu--{{ modifier }}{% endif %}">
